Implement drawing for Triangle
Also, add abstract class and start refactoring some stuff from Rectangle
to it
