+++ rectangle.py
@@ -1,7 +1,6 @@
 from PyQt5.QtCore import Qt, QPoint, QRect
 from PyQt5.QtGui import QColor
-from PyQt5.QtWidgets import QWidget
-from PyQt5 import uic
+from primitive import AbstractPrimitive
 
 NAT = 0
 NAR = 1
@@ -11,7 +10,7 @@ NX = 4
 NY = 5
 
 
-class Rectangle:
+class Rectangle(AbstractPrimitive):
     """
     Primitive figure, a rectangle
     """
@@ -56,18 +55,6 @@ class Rectangle:
         self.binds[edge] = neighbour
         fig = (self.x, self.y, self.width, self.height)
         self.modify(fig, self.mesh)
-
-    def draw(self, canvas, mesh_canvas, shift_x, shift_y, kx, ky):
-        def pixel_x(n_elements):
-            val = shift_x + self.start_x + n_elements*self.step_x
-            return int(round(val*kx))
-
-        def pixel_y(n_elements):
-            val = shift_y + self.start_y + n_elements*self.step_y
-            return int(round(val*ky))
-
-        self.draw_figure(canvas, shift_x, shift_y, kx, ky)
-        self.draw_mesh(mesh_canvas, pixel_x, pixel_y)
 
     def draw_figure(self, canvas, shift_x, shift_y, kx, ky):
         canvas.setPen(Qt.black)
@@ -390,150 +377,3 @@ class Rectangle:
                     f.write("0\n")  # Air
 
 
-class NewRectangleWidget(QWidget):
-    """
-    Важно: пользователь вводит количество узлов, но программе работать удобнее
-    с "квадратами" (по диагонали которого впоследствии образуется два элемента)
-    Так что при переходе туда-обратно придется прибавлять\отнимать единицу
-    """
-    def __init__(self):
-        QWidget.__init__(self)
-        uic.loadUi('ui/new_rectangle.ui', self)
-        self.manual_air.stateChanged.connect(self.upd_air_spinboxes)
-        self.connection_side = -1
-
-    def upd_air_spinboxes(self, value):
-        isEnabled = value == 2
-        self.air_left.setEnabled(isEnabled)
-        self.air_right.setEnabled(isEnabled)
-        self.air_bottom.setEnabled(isEnabled)
-
-    def set_data(self, rectangle):
-        self.x.setValue(rectangle.x)
-        self.y.setValue(rectangle.y)
-        self.width.setValue(rectangle.width)
-        self.height.setValue(rectangle.height)
-
-        Ntop, Nright, Nbottom, Nleft, Nx, Ny = rectangle.mesh
-        air = (Ntop, Nright, Nbottom, Nleft)
-        self.Nx.setValue(Nx+1)  # convert back from blocks to nodes
-        self.Ny.setValue(Ny+1)
-
-        dlg = [self.air_top, self.air_right, self.air_bottom, self.air_left]
-        for i in range(4):
-            dlg[i].setValue(air[i])
-            if rectangle.binds[i]:
-                dlg[i].hide()
-
-        if Nleft != Nright or Ntop != Nbottom or Nleft != Ntop:
-            self.manual_air.setChecked(True)
-
-    def get_data(self):
-        x = self.x.value()
-        y = self.y.value()
-        width = self.width.value()
-        height = self.height.value()
-        Nx = self.Nx.value() - 1  # '-1': proceed in blocks, not nodes
-        Ny = self.Ny.value() - 1
-        Ntop = self.air_top.value()  # no '-1' because 1 node belong to figure
-        Nright = self.air_right.value()
-        Nbottom = self.air_bottom.value()
-        Nleft = self.air_left.value()
-        if not self.manual_air.isChecked():  # use same air for all edges
-            if self.connection_side == 2:
-                hold = Nleft
-            else:
-                hold = Ntop
-            Ntop = hold
-            Nright = hold
-            Nbottom = hold
-            Nleft = hold
-
-        mesh = [Ntop, Nright, Nbottom, Nleft, Nx, Ny]
-        if self.connection_side != -1:
-            mesh[(self.connection_side+2) % 4] = 0
-
-        if self.connection_side == 0:
-            y -= height
-        elif self.connection_side == 3:
-            x -= width
-
-        return (x, y, width, height), mesh
-
-    def for_expanding(self, prim, side_code):
-        """
-        Расширяем фигуру, пристраивая прямоугольник со стороны side_code к prim
-
-        side_code:
-            0 - вверх
-            1 - справа
-            2 - низ
-            3 - слева
-        """
-        self.connection_side = side_code
-        pos = [self.x, self.y]
-        dim = [self.width, self.height]
-        air = [self.air_top, self.air_right, self.air_bottom, self.air_left]
-        fig = [self.Nx, self.Ny]
-        prim_pos = [prim.x, prim.y]
-        fixed_pos = [prim.y, prim.x + prim.width, prim.y + prim.height, prim.x]
-        prim_dim = [prim.width, prim.height]
-        prim_stp = [prim.step_x, prim.step_y]
-        prim_bounds = [(prim.x, prim.x + prim.width),
-                       (prim.y, prim.y + prim.height)]
-        nodes = [prim.mesh[4], prim.mesh[5]]  # mesh[4] is x
-
-        free_axis = side_code % 2  # ось, по которой можно двигать примитив
-        to_prim = (side_code+2) % 4  # сторона, с которой находится prim
-
-        pos[free_axis].setValue(prim_pos[free_axis])
-        pos[free_axis].setSingleStep(prim_stp[free_axis])
-        pos[not free_axis].setValue(fixed_pos[side_code])
-        pos[not free_axis].setEnabled(False)
-
-        air[to_prim].hide()
-
-        dim[free_axis].setValue(prim_dim[free_axis])
-        dim[free_axis].setSingleStep(prim_stp[free_axis])  # help with scale
-        dim[not free_axis].setValue(prim_stp[not free_axis])
-        dim[not free_axis].setSingleStep(prim_stp[not free_axis])
-
-        fig[free_axis].setValue(nodes[free_axis] + 1)  # +1 - go back to nodes
-        fig[free_axis].setEnabled(False)
-        self.wath_node(fig[free_axis], prim_stp[free_axis])
-        con_min, con_max = prim_bounds[free_axis]
-        self.wath_connection(pos[free_axis], con_min, con_max, free_axis == 1)
-
-        if side_code == 2:  # air_top was default, but now it disabled
-            self.air_left.setEnabled(True)
-
-    def wath_node(self, nodes, dk):
-        self.dezired_step = dk
-        if nodes == self.Nx:
-            self.width.valueChanged.connect(self.update_wathed_nodes)
-        else:
-            self.height.valueChanged.connect(self.update_wathed_nodes)
-
-    def update_wathed_nodes(self, dk):
-        if self.connection_side % 2 == 0:
-            Nk = self.Nx
-        else:
-            Nk = self.Ny
-        Nk.setValue(int(round(dk/self.dezired_step)) + 1)
-
-    def wath_connection(self, connected_side_coor, min_, max_, isVert):
-        connected_side_coor.setMaximum(max_)
-        self.allowed_min = min_
-        if isVert:
-            connected_side_coor.setMinimum(min_ - self.height.value())
-            self.height.valueChanged.connect(self.update_wathed_connection)
-        else:
-            connected_side_coor.setMinimum(min_ - self.width.value())
-            self.width.valueChanged.connect(self.update_wathed_connection)
-
-    def update_wathed_connection(self, length):
-        if self.connection_side % 2 == 1:
-            coor = self.y
-        else:
-            coor = self.x
-        coor.setMinimum(self.allowed_min - length)
