+++ ui/new_triangle.ui
@@ -189,7 +189,7 @@
     </item>
    </layout>
   </widget>
-  <widget class="QWidget" name="">
+  <widget class="QWidget" name="layoutWidget">
    <property name="geometry">
     <rect>
      <x>340</x>
@@ -202,7 +202,7 @@
     <item>
      <widget class="QLabel" name="label">
       <property name="text">
-       <string>Форма:</string>
+       <string>Вид:</string>
       </property>
       <property name="alignment">
        <set>Qt::AlignCenter</set>
